pyproject.toml
[tool.poetry]
name = "django-api-rest"
version = "0.1.0"
description = ""
authors = ["cmareau-git <charles_mareau@hotmail.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
